+++ device-bootloaders/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [SW-18460] Add more error logs for debugging.
 ### Changed
 ### Deprecated
 ### Removed